use double list instead of int lists and double arrays

+++ protocol/src/commonMain/kotlin/com/sourceplusplus/protocol/artifact/metrics/ArtifactMetrics.kt
@@ -11,5 +11,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ArtifactMetrics(
     val metricType: MetricType,
-    val values: List<Int>
+    val values: List<Double>
 )
